+++ AdBid.cpp
@@ -4,8 +4,8 @@
 
 #include "AdBid.h"
 
-AdBid::AdBid(std::string_view keyword, unsigned money) :
+AdBid::AdBid(std::string_view keyword, long money) :
   _keyword(keyword), _money(money) {}
 
-AdBidMatched::AdBidMatched(std::string_view keyword, unsigned money, const Ad* ad) :
+AdBidMatched::AdBidMatched(std::string_view keyword, long money, const Ad* ad) :
   _keyword(keyword), _money(money), _ad(ad) {}
